refactor: utils를 util로 수정
